+++ app/src/routes/editor/EditPanel/StylesTab/compound/DisplayInput.tsx
